AG-15782 Partially move shared getStyle implementation to base class
